backend/requirements.txt
flask==2.0.1
pandas>=1.3.3
google-cloud-storage==1.42.0
gunicorn==20.1.0
python-dotenv==0.19.0